fixed some bug
